Reorganized the static stuff
Now the static stuff is naturally seperate from the library stuff, the rpg.js file
is symlinked into the js folder of the static folder.  Not sure if git handles
symlinks, may need to see if there is a better way to have a file in a repo be in
two places at once.