Fixed ObjectId for database to throw error
The ObjectId throws a bson error naturally when the string is not a proper ObjectId,
the init of the database now wraps that function and throws a database error.  This
is a simple solution versus wrapping the conversion every place it occurs (the
error to be thrown will be the same as the look should fail everytime).