Updated quest completion model
The players can now request the completion of a quest, meaning that the action is no
longer restricted to DMs, but the quest is not completed until the DM approves the
request (weird verbage on this).  No tests have been written yet and the JS library
doesn't reflect this so...  Added a new endpoint to view the outstanding requests,
should probably update the status values to be more meaningful?