Expanded form handling
Forms without an action set will not be submitted and the form handler won't be
called.  If an input name has dots in it, they will be treated as child keys of the
preceding key (so foo.bar means it is setting foo: { bar: <value> } and not
foo.bar: <value>).  If no name is set on the input, it is skipped as well.