Fixed up some stuff in the js library
Now error returns the message for both native and jQuery, the user library is
updated to do this properly now.  All of the results of the all grab calls (i.e.
rpg.skill.get(), rpg.classes.get(), rpg.quest.get()) will return a keyed hash
table and not a general array.  This is still loopable and is the proper data
structure for the design of the app.  The function used to convert is exposed
through utils part of the library.