More test coverage
Using the coverage module to give test coverage for nosetests, I have tried to get
better coverage of all of my code, still trying to improve.  The goal is to be at
above 90% for all modules, then ultimately at 100% (there are a lot of dark corners
that are not hit).