Some fixes from deploying on Heroku to master
Fixed an issue with compiling the LESS into CSS, the folder it compiles to is
ignored and the compile command in Make will create the directory (unix specific for
now).  Also fixed the caching of skills in quest creation (same fix as for #18) and
improved the style appearance as with #12.