Addition of serverside caching (addresses Issue #1)
Requests that should be cached use the `@cache` decorator and requests that can
invalidate the cache use the `mark_dirty` function and pass in the path for the
requests that will be marked dirty.  This could get hairy, so probably will need
to rethink how the dirty marking is done.